+++ braid/examples/update.rs
@@ -1,0 +1,25 @@
+//! Update and engine and show a progress bar
+use braid::examples::Example;
+use braid::misc::progress_bar;
+use braid::{create_comms, EngineUpdateConfig};
+
+#[tokio::main]
+async fn main() {
+    let mut engine = Example::Animals.engine().unwrap();
+
+    let config = EngineUpdateConfig::new()
+        .default_transitions()
+        .n_iters(50)
+        .timeout(10);
+
+    let (sndr, rcvr) = create_comms();
+
+    let pbar_handle = progress_bar(engine.n_states() * config.n_iters, rcvr);
+
+    let update_handle = tokio::spawn(async move {
+        engine.update(config, Some(sndr), None).unwrap();
+    });
+
+    pbar_handle.await.expect("Failed to join progress bar");
+    update_handle.await.expect("Failed to join update");
+}