+++ braid/src/examples/mod.rs
@@ -95,10 +95,8 @@ impl Example {
         timeout: Option<u64>,
     ) -> Result<(), Error> {
         use crate::config::EngineUpdateConfig;
-        use crate::UpdateInformation;
 
         let n_states = 8;
-        let comms = UpdateInformation::new(n_states);
 
         let paths = self.paths()?;
         let codebook: Codebook = {
@@ -128,7 +126,7 @@ impl Example {
             ..Default::default()
         };
 
-        engine.update(config, Some(std::sync::Arc::new(comms)))?;
+        engine.update(config, None, None)?;
         engine.save(paths.braid.as_path(), &SaveConfig::default())?;
         Ok(())
     }